Store parameters in a list instead of a dict
While the dict is simple to implement, it lacks the ability to store
parameters with the same name multiple times. Parameter names are not
unique in all instances, with the most obvious one being a PROJ string
expressed as a pipeline, which can contain multiple '+proj'-parameters.

With this in mind a new Parameter class has been added and Operation
parameters are kept track of in a list.