Add ProjOperator
This allows access to any coordinate operation that can be done in PROJ.
This will be useful in cases where the representation of coordinates
needs to be different, for instance geodetic instead of cartesian. It
might also serve a useful purpose when either the source or target
coordinates are given as projected coordinates, but the parameter
estimation needs to be done in ECEF space. And a whole range of other
possibilites that is only limited by the imagination of the user.