Pass source and target directly to Presenter.evaluate
This separates source and target data from the results list of a
Pipeline. When evaluating the results with a Presenter it is beneficial
to get direct access to the source and target DataSource's instead of
implying them from the results list.

This refactoring effort causes changes to all existing Presenters and
number of tests. All have been adjusted accordingly.