Finally use weights in parameter estimation
HelmertTranslation modified to now determine the parameters based on
weighted averages instead of just plain averages. The Pipeline now also
delivers the weights to the operators when estimating parameters. A
bunch of minor changes made to accommodate those changes. A number of
tests have been modified and new ones added.