Implement overrides in DataSource's
With this, it is now possible to override values for sx, sy, sz, w and
t. These values will be applied for all coordinates in the DataSource.
By default they are all optional in subclasses of DataSource but can be
made mandatory for a specific subclass of DataSource, e.g. if the data
source does not include those fields natively. An example of this is
CRD-files from Bernese that does not include information of the
coordinate uncertainties.