Ensure geodetic coordinates in GeoJSON output
Coordinate.geojson_feature() now takes a transformer, so that it can
change the coordinate into geodetic coordinates. It is the
responsibility of the caller to give a suitable transformer.

This commits fixes problems in presenters that produces GeoJSON files.
As a consequence it is now mandatory to set `coordinate_type` for the
residual_presenter.