WIP: Ensure geodetic coordinates in GeoJSON output
Coordinate.geojson_feature() now takes a transformer, so that it can
change the coordinate into geodetic coordinates. It is the
responsibility of the caller to give a suitable transformer.

This commits fixes problems in presenters that produces GeoJSON files.

NOTE: Only implemented for TopocentricResidualPresenter. Needs to be
implemented for the rest of the presenters. This might require that
users input what type of coordinates they have. Needs more thought.