Make all coordinates 4D
This affects a larger chunk of the code but is mainly a matter of
extending coordinate vectors and matrices a bit and only selecting the
spatial parts of the coordinates in most cases where operations are done
on the coordinates. I bit more convoluted but essential.

Typing extending to include both CoordinateMatrix and Matrix, as well as
CoordinateVector and Vector. Some matrices and vectors contain
coordinates and can be expected to be either Nx4 or 4x1. Other matrices
and vectors can be shaped arbitrarily.