Allow non-overlapping DataSources
When processing a Pipeline DataSources are now modified so that they
only contain coordinates from the union of the source and target
stations.

When removing stations from DataSources the original ordering of the
data is comprimised. Hence Coordinate lists are sorted (in order of
stations) as a final step of DataSource instantiantion.