Implement a better check for `Operator.can_estimate()`
Instead of running the `estimate()` method we check to see if the
operator has overloaded the parent method. If it has, the child operator
should be able to estimate parameters.

This is simpler than the previous check which called the method with
dummy input, which in some cases caused issues when processing a
Pipeline.